got it to compile! it's somewhat working, but no collision/meshes and lighting still bad.
